Remove dependancy on libldap on linux
The issue is that centos 7 has the old version of openssl installed
and node12 is compiled with the new version of openssl. As libldap
on centos 7 is compiled against the old version of openssl we get a
segfault if we compile against the system version of libldap.

Thereofore I've built a version of libldap against the correct
version of openssl which i then statically link against my node
library. I've also included the ldap header file so i can remove
the dependancy on openldap-devel.

Signed-off-by: Joshua Houghton <joshua.houghton@ripjar.com>